CONCAT flag (thk frank)
behaviour has now possible flags (multiple can be used using OR (|)).
The new one introduced is CONCAT.

If an item (intermediate or final) has the CONCAT flag, the script
provided won't be used as script. Instead, it will substitute the
keyname as argument with the contents of the field "script".

For example, let's say we have a->b->c in which a has the `echo`
script, b has the CONCAT flag with the `option b` "script", and c
nothing. When c is pressed, the full executed script will be:

echo 'option b' 'c'

That is, the key name 'b' is substituted by the content of the "script".

Thanks Frank for the suggestion.